+++ mayan/apps/cabinets/events.py
@@ -8,6 +8,9 @@ namespace = EventTypeNamespace(label=_('Cabinets'), name='cabinets')
 event_cabinet_created = namespace.add_event_type(
     label=_('Cabinet created'), name='cabinet_created'
 )
+event_cabinet_deleted = namespace.add_event_type(
+    label=_('Cabinet deleted'), name='cabinet_deleted'
+)
 event_cabinet_edited = namespace.add_event_type(
     label=_('Cabinet edited'), name='cabinet_edited'
 )